Winnie helena move split (#61)
* fix import

* Pass split in get_iter

* Adapt and register readme

* Complete registration
