+++ docs/source/dataset_links/automl.rst
@@ -1,0 +1,3 @@
+Dataset Card - AutoML
+=====================================
+.. include:: ../../../src/squirrel_datasets_core/datasets/AutoMl/README.rst
