+++ docs/source/index.rst
@@ -22,6 +22,7 @@ Available Datasets
    dataset_links/imagenet
    dataset_links/monthly_german_tweets
    dataset_links/bdd100k
+   dataset_links/automl
 
 Contribute
 ------------------
